change3

+++ stats.cpp
@@ -1,5 +1,5 @@
 #include "stats.h"
 
-Stats Statistics::ComputeStatistics(const std::vector<int>& ) {
+Stats Statistics::ComputeStatistics(const std::vector<double>& ) {
     //Implement statistics here
 }
